requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
async-timeout==4.0.3
asyncio==3.4.3
bcrypt==4.1.2
black==24.4.2
build==1.2.1
certifi==2023.7.22
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorclass==2.2.2
decorator==5.1.1
distlib==0.3.7
dnspython==2.6.1
docopt==0.6.2
ecdsa==0.18.0
email_validator==2.1.1
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.2
filelock==3.13.1
flake8==7.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
identify==2.5.31
idna==3.4
iniconfig==2.0.0
ipython==8.18.0
isort==5.13.2
jedi==0.19.1
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
nodeenv==1.8.0
orjson==3.10.3
packaging==23.2
parso==0.8.3
passlib==1.7.4
pathspec==0.11.2
pexpect==4.9.0
pip-tools==7.4.1
pip-upgrader==1.4.15
platformdirs==3.11.0
pluggy==1.3.0
pre-commit==3.5.0
prompt-toolkit==3.0.41
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pycodestyle==2.11.1
pydantic==2.7.1
pydantic_core==2.18.2
pyflakes==3.2.0
Pygments==2.17.2
PyJWT==2.8.0
pyproject_hooks==1.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
rsa==4.9
setuptools==69.0.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
sqlmodel==0.0.18
stack-data==0.6.3
starlette==0.37.2
terminaltables==3.1.10
traitlets==5.13.0
typer==0.12.3
typing_extensions==4.8.0
ujson==5.9.0
urllib3==2.0.7
uvicorn==0.29.0
uvloop==0.19.0
virtualenv==20.24.6
watchfiles==0.21.0
wcwidth==0.2.12
websockets==12.0
wheel==0.43.0
